pyproject.toml
[project]
name = "usermcp"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = [
    'mcp',
    'agent',
    'llm',
    'profile',
    'user'
]
classifiers = [
    'Intended Audience :: Developers',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3'
]
dependencies = [
    "mcp[cli]>=1.12.3",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["localaplace"]
exclude = ["test", ".gitignore", "uv.lock", "dist", ".venv", ".vscode"]
